Major package restructuration.
[ac2art]

The entirety of the code is now grouped as a single package, entitled ac2art.

Contents from the previous data, neural_networks and utils modules were dispatched
following a new architecture, which aims at clearly separating stuff that users need
from those that run "behind the scene" are to be looked at when implementing additional
features or adding support for new corpora.

Docstrings, especially from the __init__.py files, were edited to give better insights
as to the package's structure, so that it is largely self-documented. Markdown readmes
are on the way - that of the corpora submodule will actually draw from the current
__init__.py docstring, which needs completion.

[models conversion]

Pre-trained model dumps will raise exception if attempted to be loaded; this can be
fixed by editing the '__class__' entry and 'architecture' 'class' sub-entries from
the dict constituting the dumps, to reflect the change of architecture.

For '__class__' the entry, replace the 'neural_networks.models' part with 'ac2art.networks'.

For each dict in the 'architecture' entry, apply the following replacements:
    'neural_networks.components.dense_layer.DenseLayer':
        'ac2art.internal.neural_layers._dense_layer.DenseLayer'
    'neural_networks.components.rnn._rnn.BidirectionalRNN':
        'ac2art.internal.neural_layers._rnn.BidirectionalRNN'
    'neural_networks.components.filters._filters.LowpassFilter':
        'ac2art.internal.neural_layers._filters._lowpass.LowpassFilter'

These modifications are easily applied by loading the dumped descriptor dict with
`np.load(<path>).tolist()` and saving back the edited dict with `np.save(<path>, <dict>)`.

[debugging]

The restructured package is supposedly debugged (e.g. `run_inversion_tasks` succeeds),
but the `corpora` submodule has not been fully tested yet and may present bugs. It will
be further tested soon, and corrected if needed.

[what's next]

* finish writing some docstrings / add readme files (see FIXME tags in the code)
* optionally go back through the network's docstring to "de-generalize" some points

* fully test (and debug if needed) ac2art.corpora (see previous section)
* add a final step to the data extraction scheme (see FIXME tag in the code)

* add a setup ruling the package's installation (optional)
  and checking dependencies (important)